Write out results
Now when you click "Init & Run" in the NEURON GUI to run the code, after running it will write out some results to results.csv. There is also an extra GUI to control whether and when to add opioids to the model cell. The actual modeling of the cell response to opioids is not yet implemented.